Fixes & Compilation flags: Fixed all compilation warnings. Added conditional compilation flags depending on build type. Fixed variable-shadowing bug (rare segfault).

+++ include/CSV.h
@@ -7,8 +7,8 @@
 
 class CSV {
     Vector2D<std::string> m_data;
-    size_t m_rows;
-    size_t m_cols;
+    long m_rows;
+    long m_cols;
 
 public:
     explicit CSV(const char* path, char separator = ',');
@@ -27,7 +27,7 @@ public:
      * \param filter function to be used in testing, returns true when value passes the test
      * \return an array containing copy of rows that pass the test.
      */
-    Vector2D<std::string> filter(int column, bool (*filter)(const std::string&value)) const;
+    Vector2D<std::string> filter(long column, bool (*filter)(const std::string&value)) const;
 
     Vector2D<std::string> filter(int column, std::string&value) const;
 };
